Update app version and refactor logging function name
Updated the application version from "1.4.4" to "1.4.5" in Dockerfile and main.py. Also, the function name for logging has been refactored from `Functions.captcha_logging` to `Functions.send_logging_message` in the `self` function in main.py. This change is reflected in two places where the function is called, once for 'verify_mass_started' and once for 'verify_mass_success'. The parameters passed to the function remain unchanged.