Updated app version and improved error handling
Updated the application version from "1.4.6" to "1.4.7" in Dockerfile and main.py. Removed the cooldown period for the 'send_pannel' command by commenting out the cooldown decorator. Enhanced error handling in the 'self' function to send a message to the user if 'verify_channel_id' is not set. Also, refined the error message for non-existent 'verify_channel' to be more specific.