Bump version to 1.5.5 and improve log handling
Updated the version number of the Discord bot from `1.5.4` to `1.5.5` in the Dockerfile and main.py. Modified the `activity_title` in the `JSONValidator` class to point to the correct GitHub repository URL. Enhanced the `log` command in the `Owner` class for better error handling and file management, including robust argument checks and efficient log file processing. Improved the logic for sending logs with streamlined zip file creation and clearer error messages for large logs.